Always ping the remote peer.
This commit unconditionally sends a ping to the remote peer every
pingInterval time (currently 2 minutes).  Additionally, if the remote
peer's protocol version is greater than btcwire.BIP0031Version, the
remote peer must send a pong in response to the ping.  A timer is
setup for such peers and will disconnect the peer if a response is
not received in pingTimeoutMinutes (currently 1 minute).